Module uic_upti_cap::pti_sniffer

Structs

ADAPTERS_MAP

AdapterItem

Address

AppInfo

MapEntry

TraceItem

UptiConfig

Enums

Modes

Constants

APP_NAME

Statics

BLE_PROTOCOL

CHANNEL_REGIONS_BAUD40_K

CHANNEL_REGIONS_BAUD100_K

CHANNEL_REGIONS_BAUD9600

CHANNEL_REGIONS_BAUD_LR

CLOSE_BRACKET

OPEN_BRACKET

OTA_PACKET_DATA

RSSI_COMPENSATION

RX_START

RX_SUCCESS

TX_START

TX_SUCCESS

ZIGBEE_PROTOCOL

ZWAVE_PROTOCOL

Functions

build_trace_message

Build TracePackage message payload

discovery_adapters

Discovers WSTK boards (adapters) in network and updates ADAPTERS_MAP content

get_appended_info

Appended Info decoder For Versions 0 and 1, APPENDED_INFO_CFG can take on these values (v bit is the version): 0b1 0 0 000 00v = 0x100 or 0x101 = Tx, SYNCWORD absent, RADIO_CONFIG absent, RSSI absent 0b1 0 0 001 00v = 0x108 or 0x109 = Tx, SYNCWORD absent, RADIO_CONFIG present, RSSI absent 0b1 0 0 010 00v = 0x110 or 0x111 = reserved 0b1 0 0 011 00v = 0x118 or 0x119 = reserved 0b1 0 0 100 00v = 0x120 or 0x121 = Tx, SYNCWORD present, RADIO_CONFIG absent, RSSI absent 0b1 0 0 101 00v = 0x128 or 0x129 = Tx, SYNCWORD present, RADIO_CONFIG present, RSSI absent 0b1 0 0 110 00v = 0x130 or 0x131 = reserved 0b1 0 0 111 00v = 0x138 or 0x139 = reserved 0b1 0 1 000 00v = 0x140 or 0x141 = reserved 0b1 0 1 001 00v = 0x148 or 0x149 = Rx, SYNCWORD absent, RADIO_CONFIG absent, RSSI present 0b1 0 1 010 00v = 0x150 or 0x151 = Rx, SYNCWORD absent, RADIO_CONFIG present, RSSI present 0b1 0 1 011 00v = 0x158 or 0x159 = reserved 0b1 0 1 100 00v = 0x160 or 0x161 = reserved 0b1 0 1 101 00v = 0x168 or 0x169 = Rx, SYNCWORD present, RADIO_CONFIG absent, RSSI present 0b1 0 1 110 00v = 0x170 or 0x171 = Rx, SYNCWORD present, RADIO_CONFIG present, RSSI present 0b1 0 1 111 00v = 0x178 or 0x179 = reserved

get_pti_data_frame

Check if frame is PTI

get_raw_frame

Identify RawFrame in TCP stream

get_supported_devices

Discovers WSTK boards with broadcast to port 19020 on all network interfaces.

on_pti_enable

Connects to adapter, creates new thread, reads trace packages from TCP stream, publishes packages to ucl/UPTIcap/<serial_number>/TracePackage topic Parsing OTA packets not implemented

parse_data

publish_list

remove_list

Removes WSTK device from ADAPTERS_MAP, stops all running threads, publishes unretain message to ucl/UPTICap/List

toggle_adapter

Controls capture sessions. Each capture session publishes data that it receives from debug channel to corresponding MQTT topic

udp_discovery